Add CurrencyService and related unit test
A new class, CurrencyService, has been introduced in the CurrencyConverter.Api project, containing a method for currency conversion. The 'UnitTest1.cs' file was removed and a new 'CurrencyConversionTest.cs' file was added to the UnitTests project to validate this new functionality. The UnitTests project now references the CurrencyConverter.Api project as well.